added various different griddings

+++ grid_parent_distributed.py
@@ -41,8 +41,8 @@ def get_and_write_ovdengrid(filepath, zoom_ncells, zoom_width, njobs, jobid):
     print("nr_cells:", nr_cells)
     print("grid cells:", ncells)
 
-    cell_bins = np.linspace(0, nr_cells, njobs + 1)
-    my_cells = np.arange(cell_bins[jobid], cell_bins[jobid + 1], 1)
+    cell_bins = np.linspace(0, nr_cells, njobs + 1, dtype=int)
+    my_cells = np.arange(cell_bins[jobid], cell_bins[jobid + 1], 1, dtype=int)
     print(my_cells)
 
     for icell in my_cells:
